All shaders fixed?
All shaders seem fixed.

+++ Assets/Shader/OLD/FirstPersonTransition.shader
@@ -1,45 +1,45 @@
-Shader "Hidden/FirstPersonTransition" {
-    Properties {
+Shader "Hidden/FirstPersonTransition" 
+{
+    Properties
+    {
         _Color ("Color", Color) = (1,1,1,1)
     }
-    SubShader { 
+    SubShader
+    {
         LOD 200
         Tags { "QUEUE"="Transparent" "RenderType"="Transparent" }
-        Pass {
+        Pass
+        {
             Tags { "QUEUE"="Transparent" "RenderType"="Transparent" }
             ZWrite Off
             Cull Off
             Fog { Mode Off }
             Blend SrcAlpha OneMinusSrcAlpha
-
             CGPROGRAM
             #pragma vertex vert
             #pragma fragment frag
-
+            #include "UnityCG.cginc"
             float4 _Color;
-
             struct appdata_t
             {
                 float4 vertex : POSITION;
             };
-
             struct v2f
             {
                 float4 vertex : POSITION;
             };
-
             v2f vert(appdata_t v)
             {
                 v2f o;
                 o.vertex = UnityObjectToClipPos(v.vertex);
                 return o;
             }
-
-            half4 frag(v2f i) : SV_TARGET
+            float4 frag(v2f i) : SV_TARGET
             {
                 return _Color;
             }
             ENDCG
         }
     }
+    
 }
